parser(comments): only allow utf-8 and (some) ascii in comments
Tested behavior both in racket and python:

- \xHH escapes in strings are UTF-8 codepoints
- \xHH escapes in bytes are bytes

This means tests that test with invalid UTF-8 may not be modeled using racket
strings.

The other issue was that decode.rkt reads the input as a string,
which (apparently) handles bad utf-8 for you: each byte becomes a replacement
character (U+FFFD), which breaks the parser's detection of bad UTF-8 as U+FFFD
is valid UTF-8.