TOML v1.0.0
Resolves duplicate tables via carrying some table metadata in a custom struct.
This enables discernment of the case where headers are duplicated but stored
values do not conflict, including cases where the conflict is with dotted keys.
Pre-commit versions required leaves to be in conflict in order to raise errors.

Also updates Unicode escapes, conforming to specifications that require `\u`
escapes to be followed by exactly 4 hex digits and and `\U` exactly 8.